% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deere-package.R
\docType{package}
\name{deere}
\alias{deere}
\alias{deere-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/deere}
\item BugReports: \url{https://gitlab.com/hrbrmstr/deere/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
