% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_coredata_ts}
\alias{from_coredata_ts}
\title{Convert timestampes from Apple "CoreData" format to something usable}
\usage{
from_coredata_ts(x, tz = NULL)
}
\arguments{
\item{x}{timestamps in Apple "CoreData" format(dates or times)}

\item{tz}{passed on to the convertion to a \code{POSIXct} object. Def: \code{NULL}.}
}
\description{
Convert timestampes from Apple "CoreData" format to something usable
}
