% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-backups.R
\name{list_ios_backups}
\alias{list_ios_backups}
\alias{platform_ios_backup_dir}
\title{List iOS backups available on this system}
\usage{
list_ios_backups(dir = platform_ios_backup_dir())

platform_ios_backup_dir()
}
\arguments{
\item{dir}{backup dir (will attempt to be auto-derived)}
}
\description{
List iOS backups available on this system
}
\examples{
list_ios_backups()
}
