% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_mowerplus.R
\name{src_mowerplus}
\alias{src_mowerplus}
\title{Find and sync a copy of the latest MowerPlus database file from an iOS backup}
\usage{
src_mowerplus(backup_id, data_loc = "~/Data", overwrite = TRUE)
}
\arguments{
\item{backup_id}{the giant hex string of a folder name}

\item{data_loc}{where \code{mowtrack.sqlite} will be sync'd}

\item{overwrite}{nuke ^^ if present (def: \code{TRUE})}
}
\description{
Find and sync a copy of the latest MowerPlus database file from an iOS backup
}
\note{
You may need to \href{https://rud.is/b/2019/06/02/trawling-through-ios-backups-for-treasure-a-k-a-how-to-fish-for-target-files-in-ios-backups-with-r/}{setup permissions}
to be able to use this method depending on which macOS version you're on.
}
\examples{
\dontrun{
mow_db <- src_mowerplus("28500cd31b9580aaf5815c695ebd3ea5f7455628")

mow_db

glimpse(tbl(mow_db, "ZMOWER"))

glimpse(tbl(mow_db, "ZACTIVITY"))

}
}
